requirements.txt
ansible-vault
pyyaml
